modules/04-portal/portal-api/app/requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
kubernetes==30.1.0
psycopg2-binary==2.9.9
requests==2.32.3
